Remove reference package dependency from the api.
